Extract QuPathGUI classes
A step towards shrinking and simplifying QuPathGUI
