+++ qupath-gui-fx/src/main/java/qupath/lib/gui/QuPathGUI.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2022 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2023 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -24,9 +24,6 @@
 package qupath.lib.gui;
 
 import java.awt.Desktop;
-import java.awt.Shape;
-import java.awt.geom.AffineTransform;
-import java.awt.geom.Area;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -36,7 +33,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.lang.Thread.UncaughtExceptionHandler;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URLDecoder;
@@ -74,13 +70,13 @@ import javax.script.ScriptException;
 import javax.swing.SwingUtilities;
 
 import org.controlsfx.control.action.Action;
-import org.controlsfx.control.action.ActionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 
+import javafx.application.Application;
 import javafx.application.HostServices;
 import javafx.application.Platform;
 import javafx.beans.binding.Bindings;
@@ -101,7 +97,6 @@ import javafx.collections.ObservableList;
 import javafx.embed.swing.JFXPanel;
 import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.geometry.Insets;
 import javafx.geometry.Orientation;
 import javafx.geometry.Pos;
@@ -113,18 +108,14 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.ButtonBar.ButtonData;
 import javafx.scene.control.ButtonType;
-import javafx.scene.control.CheckMenuItem;
 import javafx.scene.control.ComboBox;
-import javafx.scene.control.ContextMenu;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
 import javafx.scene.control.MenuItem;
-import javafx.scene.control.RadioMenuItem;
 import javafx.scene.control.SeparatorMenuItem;
 import javafx.scene.control.SplitPane;
-import javafx.scene.control.SplitPane.Divider;
 import javafx.scene.control.Tab;
 import javafx.scene.control.TabPane;
 import javafx.scene.control.TabPane.TabClosingPolicy;
@@ -134,12 +125,10 @@ import javafx.scene.control.TableRow;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextInputControl;
 import javafx.scene.control.TitledPane;
-import javafx.scene.control.ToggleGroup;
 import javafx.scene.control.ToolBar;
 import javafx.scene.control.Tooltip;
 import javafx.scene.control.TreeTableView;
 import javafx.scene.control.TreeView;
-import javafx.scene.effect.DropShadow;
 import javafx.scene.image.Image;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
@@ -147,24 +136,13 @@ import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.input.MouseButton;
-import javafx.scene.input.RotateEvent;
-import javafx.scene.input.ScrollEvent;
-import javafx.scene.input.ZoomEvent;
-import javafx.scene.layout.Border;
 import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.BorderStroke;
-import javafx.scene.layout.BorderStrokeStyle;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Region;
-import javafx.scene.paint.Color;
-import javafx.scene.shape.Ellipse;
-import javafx.scene.shape.Rectangle;
 import javafx.stage.Screen;
 import javafx.stage.Stage;
 import javafx.stage.Window;
 import javafx.stage.WindowEvent;
-import javafx.util.Duration;
-import jfxtras.scene.menu.CirclePopupMenu;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.common.ThreadTools;
 import qupath.lib.common.Timeit;
@@ -206,7 +184,6 @@ import qupath.lib.gui.scripting.ScriptEditor;
 import qupath.lib.gui.scripting.ScriptEditorControl;
 import qupath.lib.gui.scripting.languages.GroovyLanguage;
 import qupath.lib.gui.scripting.languages.ScriptLanguageProvider;
-import qupath.lib.gui.tools.ColorToolsFX;
 import qupath.lib.gui.tools.CommandFinderTools;
 import qupath.lib.gui.tools.GuiTools;
 import qupath.lib.gui.tools.MenuTools;
@@ -215,9 +192,7 @@ import qupath.lib.gui.tools.IconFactory.PathIcons;
 import qupath.lib.gui.viewer.DragDropImportListener;
 import qupath.lib.gui.viewer.OverlayOptions;
 import qupath.lib.gui.viewer.QuPathViewer;
-import qupath.lib.gui.viewer.QuPathViewerListener;
 import qupath.lib.gui.viewer.QuPathViewerPlus;
-import qupath.lib.gui.viewer.ViewerPlusDisplayOptions;
 import qupath.lib.gui.viewer.OverlayOptions.DetectionDisplayMode;
 import qupath.lib.gui.viewer.tools.PathTool;
 import qupath.lib.gui.viewer.tools.PathTools;
@@ -231,14 +206,7 @@ import qupath.lib.images.servers.ImageServerProvider;
 import qupath.lib.images.servers.ImageServers;
 import qupath.lib.images.servers.ServerTools;
 import qupath.lib.io.PathIO;
-import qupath.lib.objects.PathAnnotationObject;
-import qupath.lib.objects.PathObject;
-import qupath.lib.objects.PathObjectTools;
-import qupath.lib.objects.PathObjects;
-import qupath.lib.objects.TMACoreObject;
 import qupath.lib.objects.classes.PathClass;
-import qupath.lib.objects.hierarchy.PathObjectHierarchy;
-import qupath.lib.objects.hierarchy.TMAGrid;
 import qupath.lib.plugins.PathInteractivePlugin;
 import qupath.lib.plugins.PathPlugin;
 import qupath.lib.plugins.parameters.ParameterList;
@@ -246,8 +214,6 @@ import qupath.lib.projects.Project;
 import qupath.lib.projects.ProjectIO;
 import qupath.lib.projects.ProjectImageEntry;
 import qupath.lib.projects.Projects;
-import qupath.lib.roi.RoiTools;
-import qupath.lib.roi.interfaces.ROI;
 import qupath.lib.scripting.QP;
 import qupath.lib.scripting.ScriptParameters;
 import qupath.lib.scripting.languages.ExecutableLanguage;
@@ -294,8 +260,6 @@ public class QuPathGUI {
 	 */
 	public static final int TOOLBAR_ICON_SIZE = 16;
 
-	MultiviewManager viewerManager;
-	
 	private ObjectProperty<Project<BufferedImage>> projectProperty = new SimpleObjectProperty<>();
 	
 	private ProjectBrowser projectBrowser;
@@ -305,21 +269,12 @@ public class QuPathGUI {
 	 */
 	private PreferencePane prefsPane;
 	
-	/**
-	 * The current ImageData in the current QuPathViewer
-	 */
-	private ObjectProperty<ImageData<BufferedImage>> imageDataProperty = new SimpleObjectProperty<>();
-	
 	// Initializing the MenuBar here caused some major trouble (including segfaults) in OSX...
 	private MenuBar menuBar;
-
-	private BooleanProperty zoomToFit = new SimpleBooleanProperty(false);
 	
 	private BorderPane pane; // Main component, to hold toolbar & splitpane
 	private TabPane analysisTabPane = new TabPane();
 	private Region mainViewerPane;
-	
-	private ViewerPlusDisplayOptions viewerDisplayOptions = new ViewerPlusDisplayOptions();
 	
 	/**
 	 * Default options used for viewers
@@ -329,7 +284,13 @@ public class QuPathGUI {
 	/**
 	 * Default region store used by viewers for tile caching and repainting
 	 */
-	private DefaultImageRegionStore imageRegionStore;
+	private DefaultImageRegionStore imageRegionStore = ImageRegionStoreFactory.createImageRegionStore(QuPathGUI.getTileCacheSizeBytes());
+	
+	private MultiviewManager viewerManager = new MultiviewManager(this);
+	
+	MultiviewManager getViewerManager() {
+		return viewerManager;
+	}
 
 	private ToolBarComponent toolbar; // Top component
 	private SplitPane splitPane = new SplitPane(); // Main component
@@ -338,7 +299,7 @@ public class QuPathGUI {
 	
 	private Stage stage;
 	
-	private boolean isStandalone = false;
+	private boolean isStandalone = true;
 	
 	private DragDropImportListener dragAndDrop = new DragDropImportListener(this);
 	
@@ -355,7 +316,7 @@ public class QuPathGUI {
 	
 	private BooleanBinding noProject = projectProperty.isNull();
 	private BooleanBinding noViewer = viewerProperty.isNull();
-	private BooleanBinding noImageData = imageDataProperty.isNull();
+	private BooleanBinding noImageData = imageDataProperty().isNull();
 	
 	private BooleanProperty pluginRunning = new SimpleBooleanProperty(false);
 	private BooleanProperty scriptRunning = new SimpleBooleanProperty(false);
@@ -425,7 +386,7 @@ public class QuPathGUI {
 	}
 	
 	// TODO: Remove this command whenever annotations can be applied more easily
-	private Action createImageDataAction(Consumer<ImageData<BufferedImage>> command, String name) {
+	Action createImageDataAction(Consumer<ImageData<BufferedImage>> command, String name) {
 		var action = createImageDataAction(command);
 		action.setText(name);
 		return action;
@@ -506,7 +467,7 @@ public class QuPathGUI {
 		 * Apply 'zoom-to-fit' setting to all viewers
 		 */
 		@ActionIcon(PathIcons.ZOOM_TO_FIT)
-		public final Action ZOOM_TO_FIT = ActionTools.createSelectableAction(zoomToFit, "Zoom to fit");
+		public final Action ZOOM_TO_FIT = ActionTools.createSelectableAction(viewerManager.zoomToFitProperty(), "Zoom to fit");
 		
 		// Tool actions
 		/**
@@ -581,19 +542,19 @@ public class QuPathGUI {
 		 */
 		@ActionIcon(PathIcons.OVERVIEW)
 		@ActionDescription("Show/hide overview image (top right)")
-		public final Action SHOW_OVERVIEW = ActionTools.createSelectableAction(viewerDisplayOptions.showOverviewProperty(), "Show slide overview");
+		public final Action SHOW_OVERVIEW = ActionTools.createSelectableAction(viewerManager.showOverviewProperty(), "Show slide overview");
 		/**
 		 * Toggle the cursor location display on the viewers.
 		 */
 		@ActionIcon(PathIcons.LOCATION)
 		@ActionDescription("Show/hide location text (bottom right)")
-		public final Action SHOW_LOCATION = ActionTools.createSelectableAction(viewerDisplayOptions.showLocationProperty(), "Show cursor location");
+		public final Action SHOW_LOCATION = ActionTools.createSelectableAction(viewerManager.showLocationProperty(), "Show cursor location");
 		/**
 		 * Toggle the scalebar display on the viewers.
 		 */
 		@ActionIcon(PathIcons.SHOW_SCALEBAR)
 		@ActionDescription("Show/hide scalebar (bottom left)")
-		public final Action SHOW_SCALEBAR = ActionTools.createSelectableAction(viewerDisplayOptions.showScalebarProperty(), "Show scalebar");
+		public final Action SHOW_SCALEBAR = ActionTools.createSelectableAction(viewerManager.showScalebarProperty(), "Show scalebar");
 		/**
 		 * Toggle the counting grid display on the viewers.
 		 */
@@ -623,7 +584,7 @@ public class QuPathGUI {
 		/**
 		 * Add a note to any selected TMA core.
 		 */
-		private final Action TMA_ADD_NOTE = createImageDataAction(imageData -> TMACommands.promptToAddNoteToSelectedCores(imageData), "Add TMA note");
+		public final Action TMA_ADD_NOTE = createImageDataAction(imageData -> TMACommands.promptToAddNoteToSelectedCores(imageData), "Add TMA note");
 		
 		// Overlay options actions
 		/**
@@ -829,24 +790,14 @@ public class QuPathGUI {
 	
 	
 	/**
-	 * Create a QuPath instance, optionally initializing it with a path to open.
-	 * <p>
-	 * It is also possible to specify that QuPath runs as a standalone application or not.
-	 * The practical difference is that, if a standalone application, QuPath may call System.exit(0)
-	 * when its window is closed; otherwise, it must not for fear or bringing the host application with it.
-	 * <p>
-	 * If QuPath is launched, for example, from a running Fiji instance then isStandalone should be false.
+	 * Create a new QuPath instance.
 	 * 
-	 * @param services host services available during startup; may be null, but required for some functionality (e.g. opening a webpage in the host browser)
 	 * @param stage a stage to use for the main QuPath window
-	 * @param isStandalone true if QuPath should be run as a standalone application
 	 */
-	private QuPathGUI(final HostServices services, final Stage stage, final boolean isStandalone) {
+	private QuPathGUI(final Stage stage) {
 		super();
 		
 		var timeit = new Timeit().start("Starting");
-		
-		this.hostServices = services;
 		
 		if (PathPrefs.doCreateLogFilesProperty().get()) {
 			File fileLogging = tryToStartLogFile();
@@ -870,8 +821,6 @@ public class QuPathGUI {
 		
 		// Set up cache
 		timeit.checkpoint("Creating tile cache");
-		imageRegionStore = ImageRegionStoreFactory.createImageRegionStore(QuPathGUI.getTileCacheSizeBytes());
-		
 		PathPrefs.tileCachePercentageProperty().addListener((v, o, n) -> {
 			imageRegionStore.getCache().clear();
 		});
@@ -879,10 +828,9 @@ public class QuPathGUI {
 		ImageServerProvider.setCache(imageRegionStore.getCache(), BufferedImage.class);
 		
 		this.stage = stage;
-		this.isStandalone = isStandalone;
 		
 		timeit.checkpoint("Creating menus");
-
+		
 		menuBar = new MenuBar(
 				Arrays.asList("File", "Edit", "Tools", "View", "Objects", "TMA", "Measure", "Automate", "Analyze", "Classify", "Extensions", "Help")
 				.stream().map(Menu::new).toArray(Menu[]::new)
@@ -907,10 +855,6 @@ public class QuPathGUI {
 		timeit.checkpoint("Creating preferences");
 		prefsPane = new PreferencePane();
 		
-		// Activate the log at an early stage
-		// TODO: NEED TO TURN ON LOG!
-//		Action actionLog = createAction(GUIActions.SHOW_LOG);
-		
 		// Turn off the use of ImageIODiskCache (it causes some trouble)
 		ImageIO.setUseCache(false);
 		
@@ -919,18 +863,12 @@ public class QuPathGUI {
 		
 		logger.trace("Time to tools: {} ms", (System.currentTimeMillis() - startTime));
 				
-		// Initialize all tools
-//		initializeTools();
-
-		// Initialize main GUI
-//		initializeMainComponent();
-		
 		// Set this as the current instance
 		if (instance == null || instance.getStage() == null || !instance.getStage().isShowing())
 			instance = this;
 		
 		// Ensure the user is notified of any errors from now on
-		Thread.setDefaultUncaughtExceptionHandler(new QuPathUncaughtExceptionHandler());
+		Thread.setDefaultUncaughtExceptionHandler(new QuPathUncaughtExceptionHandler(this));
 		
 		
 		logger.trace("Time to main component: {} ms", (System.currentTimeMillis() - startTime));
@@ -998,8 +936,6 @@ public class QuPathGUI {
 		timeit.checkpoint("Setting icons");
 		stage.getIcons().addAll(loadIconList());
 		
-		// TODO: Make the script menus also available within the script editor itself
-		
 		// Add scripts menu (delayed to here, since it takes a bit longer)
 		timeit.checkpoint("Adding script menus");
 		Menu menuAutomate = getMenu("Automate", false);
@@ -1060,7 +996,7 @@ public class QuPathGUI {
 		logger.debug("Startup time: {} ms", (endTime - startTime));
 				
 		// Add listeners to set default project and image data
-		imageDataProperty.addListener((v, o, n) -> QP.setDefaultImageData(n));
+		imageDataProperty().addListener((v, o, n) -> QP.setDefaultImageData(n));
 		projectProperty.addListener((v, o, n) -> QP.setDefaultProject(n));
 		
 		// Run startup script, if we can
@@ -1326,38 +1262,37 @@ public class QuPathGUI {
 	/**
 	 * Static method to launch QuPath on the JavaFX Application thread.
 	 * <p>
-	 * This can potentially be used from other environments (e.g. MATLAB, Fiji, Python).
-	 * It is assumed that it is being launched from a JavaFX (without {@link HostServices} available) or AWT/Swing application; 
-	 * if this is not the case, use {@link #launchQuPath(HostServices, boolean)} instead.
+	 * This is typically expected to be called in an {@link Application#start(Stage)} method, 
+	 * although could potentially be called later from another JavaFX application.
 	 * <p>
-	 * Afterwards, calls to {@link #getInstance()} will return the QuPath instance as soon as it is available.
-	 * However, note that depending upon the thread from which this method is called, the QuPath instance may <i>not</i> 
-	 * be available until some time after the method returns.
+	 * Afterwards, calls to {@link #getInstance()} will return the QuPath instance.
 	 * <p>
 	 * If there is already an instance of QuPath running, this ensures that it is visible - but otherwise does nothing.
+	 * 
+	 * @param stage the stage to use for the QuPath UI (usually from {@link Application#start(Stage)})
+	 * @return
+	 * @throws IllegalStateException if an instance of QuPath is already running (i.e. {@link #getInstance()} is not null)
+	 * @see #launchInstanceFromSwing()
 	 */
-	public static void launchQuPath() {
-		launchQuPath(null, true);
+	public static QuPathGUI createInstance(Stage stage) throws IllegalStateException {
+		QuPathGUI instance = getInstance();
+		if (instance != null) {
+			throw new IllegalStateException("QuPathGUI already exists, cannot create a new instance!");
+		}
+		if (stage == null)
+			stage = new Stage();
+		return new QuPathGUI(stage);
 	}
 	
 	/**
-	 * Launch QuPath from an AWT/Swing application.
-	 * Equivalent to calling {@code #launchQuPath(hostServices, true)}
-	 * @param hostServices
-	 * @deprecated as of v0.2.3 in favor of {@link #launchQuPath(HostServices, boolean)}
-	 */
-	@Deprecated
-	public static void launchQuPath(HostServices hostServices) {
-		launchQuPath(hostServices, true);
-	}
-	
-	
-	public static QuPathGUI launchQuPathStandalone(HostServices hostServices, Stage stage) {
-		return new QuPathGUI(hostServices, stage, true);
-	}
-	
-	/**
-	 * Static method to launch QuPath on the JavaFX Application thread.
+	 * Static method to launch QuPath from within a Swing/AWT application.
+	 * <p>
+	 * This aims to handle several things:
+	 * <ul>
+	 *   <li>initializing JavaFX in the appropriate thread</li>
+	 *   <li>flagging that shutting down QuPath should not terminate the JVM</li>
+	 *   <li>showing the QuPath UI window</li>
+	 * </ul>
 	 * <p>
 	 * This can potentially be used from other environments (e.g. MATLAB, Fiji, Python).
 	 * Afterwards, calls to {@link #getInstance()} will return the QuPath instance as soon as it is available.
@@ -1365,63 +1300,49 @@ public class QuPathGUI {
 	 * be available until some time after the method returns.
 	 * <p>
 	 * If there is already an instance of QuPath running, this requests that it is made visible - but otherwise does nothing.
-	 * 
-	 * @param hostServices JavaFX HostServices if available, otherwise null
-	 * @param isSwing if true, it is assumed that the launch is being requested from another Swing-based Java application.
-	 *                This results in an alternative method of starting the JavaFX runtime, which may be more reliable. 
-	 *                However, when called from a non-swing app in some cases the thread may freeze.
 	 */
-	public static void launchQuPath(HostServices hostServices, boolean isSwing) {
+	public static void launchInstanceFromSwing() {
 		
-		QuPathGUI instance = getInstance();
-		if (instance != null) {
-			logger.info("Request to launch QuPath - will try to show existing instance instead");
-			if (Platform.isFxApplicationThread())
-				instance.getStage().show();
-			else {
-				Platform.runLater(() -> instance.getStage().show());
-			}
-			return;
-		}
+		// If we are starting from a Swing application, try to ensure we are on the correct thread
+		// (This can be particularly important on macOS)
+		if (Platform.isFxApplicationThread() || SwingUtilities.isEventDispatchThread()) {
+			launchNonStandaloneInstance();
+		} else {
+			System.out.println("QuPath launch requested in " + Thread.currentThread());
+			SwingUtilities.invokeLater(() -> launchNonStandaloneInstance());
+			// Required to be able to restart QuPath... or probably any JavaFX application
+			Platform.setImplicitExit(false);
+		}		
+		
+	}
+		
+	private static void launchNonStandaloneInstance() {
 		
 		if (Platform.isFxApplicationThread()) {
 			System.out.println("Launching new QuPath instance...");
 			logger.info("Launching new QuPath instance...");
-			Stage stage = new Stage();
-			QuPathGUI qupath = new QuPathGUI(hostServices, stage, false);
+			QuPathGUI qupath = createInstance(new Stage());
+			qupath.isStandalone = false;
 			qupath.getStage().show();
-			System.out.println("Done!");
-			return;
-		}
-		
-		System.out.println("QuPath launch requested in " + Thread.currentThread());
-		
-		if (isSwing) {
-			// If we are starting from a Swing application, try to ensure we are on the correct thread
-			// (This can be particularly important on macOS)
-			if (SwingUtilities.isEventDispatchThread()) {
-				System.out.println("Initializing with JFXPanel...");
-				new JFXPanel(); // To initialize
-				Platform.runLater(() -> launchQuPath(hostServices, true));
-				return;
-			} else {
-				SwingUtilities.invokeLater(() -> launchQuPath(hostServices, true));
-				// Required to be able to restart QuPath... or probably any JavaFX application
-				Platform.setImplicitExit(false);
-			}
+			System.out.println("Done!");			
+		} else if (SwingUtilities.isEventDispatchThread()) {
+			System.out.println("Initializing with JFXPanel...");
+			new JFXPanel(); // To initialize
+			Platform.runLater(() -> launchNonStandaloneInstance());
 		} else {
 			try {
 				// This will fail if already started... but unfortunately there is no method to query if this is the case
 				System.out.println("Calling Platform.startup()...");
-				Platform.startup(() -> launchQuPath(hostServices, false));
-			} catch (Exception e) {
+				Platform.startup(() -> launchNonStandaloneInstance());
+			} catch (IllegalStateException e) {
 				System.err.println("If JavaFX is initialized, be sure to call launchQuPath() on the Application thread!");
 				System.out.println("Calling Platform.runLater()...");
-				Platform.runLater(() -> launchQuPath(hostServices, false));
+				Platform.runLater(() -> launchNonStandaloneInstance());
 			}
 		}
 		
 	}
+	
 	
 	
 	
@@ -1450,6 +1371,27 @@ public class QuPathGUI {
 			Dialogs.showErrorMessage("Show URL", "Sorry, unable to launch a browser to open \n" + url);
 			return false;
 		}
+	}
+	
+	/**
+	 * Set the {@link HostServices} for this instance of QuPath.
+	 * These can be provided from a standalone JavaFX {@link Application}.
+	 * Making them available means they can be used e.g. to launch a browser window.
+	 * 
+	 * @param services
+	 * @see #getHostServices()
+	 */
+	void setHostServices(HostServices services) {
+		this.hostServices = services;
+	}
+	
+	/**
+	 * Get the {@link HostServices}, or null if none have been set.
+	 * @return
+	 * @see #setHostServices(HostServices)
+	 */
+	HostServices getHostServices() {
+		return this.hostServices;
 	}
 	
 	
@@ -1985,7 +1927,7 @@ public class QuPathGUI {
 		availablePathClasses = FXCollections.observableArrayList();
 		Set<PathClass> pathClasses = new LinkedHashSet<>();		
 		try {
-			pathClasses.addAll(loadPathClasses());			
+			pathClasses.addAll(loadPathClassesFromPrefs());			
 		} catch (Exception e) {
 			logger.error("Unable to load PathClasses", e);
 		}
@@ -2050,7 +1992,7 @@ public class QuPathGUI {
 	 * 
 	 * @return
 	 */
-	private static List<PathClass> loadPathClasses() {
+	private static List<PathClass> loadPathClassesFromPrefs() {
 		byte[] bytes = PathPrefs.getUserPreferences().getByteArray("defaultPathClasses", null);
 		if (bytes == null || bytes.length == 0)
 			return Collections.emptyList();
@@ -2151,14 +2093,6 @@ public class QuPathGUI {
 		
 		pane = new BorderPane();
 		
-		// Create a reasonably-sized viewer
-		QuPathViewerPlus viewer = new QuPathViewerPlus(null, imageRegionStore, overlayOptions, viewerDisplayOptions);
-		
-		
-
-		// Add analysis panel & viewer to split pane
-		viewerManager = new MultiviewManager(viewer);
-		
 		viewerProperty.bind(viewerManager.activeViewerProperty());
 		
 		// Now that we have a viewer, we can create an undo/redo manager
@@ -2196,13 +2130,6 @@ public class QuPathGUI {
 		setAnalysisPaneVisible(showAnalysisPane.get());
 		showAnalysisPane.addListener((v, o, n) -> setAnalysisPaneVisible(n));
 		
-//		setInitialLocationAndMagnification(getViewer());
-
-		// Prepare the viewer
-		setupViewer(viewerManager.getActiveViewer());
-
-		// Ensure the mode is set
-
 		// Ensure actions are set appropriately
 		selectedToolProperty.addListener((v, o, n) -> {
 			Action action = toolActions.get(n);
@@ -2214,7 +2141,7 @@ public class QuPathGUI {
 			if (n == PathTools.POINTS)
 				defaultActions.COUNTING_PANEL.handle(null);
 			
-			updateCursor();			
+			updateCursorForSelectedTool();			
 		});
 		setSelectedTool(getSelectedTool());
 		
@@ -2316,12 +2243,6 @@ public class QuPathGUI {
 //		logger.debug("Tools activated for {}", viewer);
 	}
 	
-	
-	private void deactivateTools(final QuPathViewerPlus viewer) {
-		viewer.setActiveTool(null);
-	}
-	
-	
 	/**
 	 * Get the {@link ClassLoader} used to load extensions.
 	 * @return
@@ -2342,534 +2263,6 @@ public class QuPathGUI {
 	
 	
 	
-	void setupViewer(final QuPathViewerPlus viewer) {
-		
-		viewer.getView().setFocusTraversable(true);
-		
-		// Update active viewer as required
-		viewer.getView().focusedProperty().addListener((e, f, nowFocussed) -> {
-			if (nowFocussed) {
-				viewerManager.setActiveViewer(viewer);
-			}
-		});
-		
-		viewer.getView().addEventFilter(MouseEvent.MOUSE_PRESSED, e -> viewer.getView().requestFocus());
-
-		viewer.zoomToFitProperty().bind(zoomToFit);
-		
-		// Create popup menu
-		setViewerPopupMenu(viewer);
-		
-		
-		viewer.getView().widthProperty().addListener((e, f, g) -> {
-			if (viewer.getZoomToFit())
-				updateMagnificationString();
-		});
-		viewer.getView().heightProperty().addListener((e, f, g) -> {
-			if (viewer.getZoomToFit())
-				updateMagnificationString();
-		});
-		
-		// Enable drag and drop
-		dragAndDrop.setupTarget(viewer.getView());
-		
-		
-		
-		
-		// Listen to the scroll wheel
-		viewer.getView().setOnScroll(e -> {
-			if (viewer == viewerManager.getActiveViewer() || !viewerManager.getSynchronizeViewers()) {
-				double scrollUnits = e.getDeltaY() * PathPrefs.getScaledScrollSpeed();
-				
-				// Use shift down to adjust opacity
-				if (e.isShortcutDown()) {
-					OverlayOptions options = viewer.getOverlayOptions();
-					options.setOpacity((float)(options.getOpacity() + scrollUnits * 0.001));
-					return;
-				}
-				
-				// Avoid zooming at the end of a gesture when using touchscreens
-				if (e.isInertia())
-					return;
-				
-				if (PathPrefs.invertScrollingProperty().get())
-					scrollUnits = -scrollUnits;
-				double newDownsampleFactor = viewer.getDownsampleFactor() * Math.pow(viewer.getDefaultZoomFactor(), scrollUnits);
-				newDownsampleFactor = Math.min(viewer.getMaxDownsample(), Math.max(newDownsampleFactor, viewer.getMinDownsample()));
-				viewer.setDownsampleFactor(newDownsampleFactor, e.getX(), e.getY());
-			}
-		});
-		
-		
-		viewer.getView().addEventFilter(RotateEvent.ANY, e -> {
-			if (!PathPrefs.useRotateGesturesProperty().get())
-				return;
-//			logger.debug("Rotating: " + e.getAngle());
-			viewer.setRotation(viewer.getRotation() + Math.toRadians(e.getAngle()));
-			e.consume();
-		});
-
-		viewer.getView().addEventFilter(ZoomEvent.ANY, e -> {
-			if (!PathPrefs.useZoomGesturesProperty().get())
-				return;
-			double zoomFactor = e.getZoomFactor();
-			if (Double.isNaN(zoomFactor))
-				return;
-			
-			logger.debug("Zooming: " + e.getZoomFactor() + " (" + e.getTotalZoomFactor() + ")");
-			viewer.setDownsampleFactor(viewer.getDownsampleFactor() / zoomFactor, e.getX(), e.getY());
-			e.consume();
-		});
-		
-		viewer.getView().addEventFilter(ScrollEvent.ANY, new ScrollEventPanningFilter(viewer));
-		
-		
-		viewer.getView().addEventHandler(KeyEvent.KEY_PRESSED, e -> {
-			PathObject pathObject = viewer.getSelectedObject();
-			if (!e.isConsumed() && pathObject != null) {
-				if (pathObject.isTMACore()) {
-					TMACoreObject core = (TMACoreObject)pathObject;
-					if (e.getCode() == KeyCode.ENTER) {
-						defaultActions.TMA_ADD_NOTE.handle(new ActionEvent(e.getSource(), e.getTarget()));
-						e.consume();
-					} else if (e.getCode() == KeyCode.BACK_SPACE) {
-						core.setMissing(!core.isMissing());
-						viewer.getHierarchy().fireObjectsChangedEvent(this, Collections.singleton(core));
-						e.consume();
-					}
-				} else if (pathObject.isAnnotation()) {
-					if (e.getCode() == KeyCode.ENTER) {
-						GuiTools.promptToSetActiveAnnotationProperties(viewer.getHierarchy());
-						e.consume();
-					}
-				}
-			}
-		});
-		
-
-	}
-	
-	
-	
-	
-	static class ScrollEventPanningFilter implements EventHandler<ScrollEvent> {
-		
-		private QuPathViewer viewer;
-		private boolean lastTouchEvent = false;
-		private double deltaX = 0;
-		private double deltaY = 0;
-		private long lastTimestamp = 0L;
-		
-		ScrollEventPanningFilter(final QuPathViewer viewer) {
-			this.viewer = viewer;
-		}
-
-		@Override
-		public void handle(ScrollEvent e) {
-			// Check if we'd rather be using scroll to do something else (e.g. zoom, adjust opacity)
-			boolean wouldRatherDoSomethingElse = e.getTouchCount() == 0 && (!PathPrefs.useScrollGesturesProperty().get() || e.isShiftDown() || e.isShortcutDown());
-			if (wouldRatherDoSomethingElse) {
-				return;
-			}
-			
-			if (viewer.getZoomToFit())
-				return;
-			
-			// Don't pan with inertia events (use the 'mover' instead)
-			if (e.isInertia()) {
-				e.consume();
-				return;
-			}
-			
-			// Return if we aren't using a touchscreen, and we don't want to handle scroll gestures - 
-			// but don't consume the event so that it can be handled elsewhere
-			lastTouchEvent = e.getTouchCount() != 0;
-			if (!lastTouchEvent && !PathPrefs.useScrollGesturesProperty().get() || e.isShiftDown() || e.isShortcutDown()) {
-				return;
-			}
-			// Swallow the event if we're using a touch screen without the move tool selected - we want to draw instead
-			if (lastTouchEvent && viewer.getActiveTool() != PathTools.MOVE) {
-				e.consume();
-				return;
-			}
-			
-//			// If this is a SCROLL_FINISHED event, continue moving with the last starting velocity - but ignore inertia
-			if (!lastTouchEvent && e.getEventType() == ScrollEvent.SCROLL_FINISHED) {
-				if (System.currentTimeMillis() - lastTimestamp < 100L) {
-					viewer.requestStartMoving(deltaX, deltaY);
-					viewer.requestDecelerate();					
-				}
-				deltaX = 0;
-				deltaY = 0;
-				e.consume();
-				return;
-			}
-//			viewer.requestStopMoving();
-			
-			// Use downsample since shift will be defined in full-resolution pixel coordinates
-			double dx = e.getDeltaX() * viewer.getDownsampleFactor();
-			double dy = e.getDeltaY() * viewer.getDownsampleFactor();
-			
-			// When e.isInertia() == TRUE on OSX, the results are quite annoyingly 'choppy' - x,y values are often passed separately
-//			System.err.println(String.format("dx=%.1f, dy=%.1f %s", e.getDeltaX(), e.getDeltaY(), (e.isInertia() ? "-Inertia" : "")));
-			
-			// Flip scrolling direction if necessary
-			if (PathPrefs.invertScrollingProperty().get()) {
-				dx = -dx;
-				dy = -dy;
-			}
-			
-			// Handle rotation
-			if (viewer.isRotated()) {
-				double cosTheta = Math.cos(-viewer.getRotation());
-				double sinTheta = Math.sin(-viewer.getRotation());
-				double dx2 = cosTheta*dx - sinTheta*dy;
-				double dy2 = sinTheta*dx + cosTheta*dy;
-				dx = dx2;
-				dy = dy2;
-			}
-
-			// Shift the viewer
-			viewer.setCenterPixelLocation(
-					viewer.getCenterPixelX() - dx,
-					viewer.getCenterPixelY() - dy);
-			
-			// Retain deltas in case we need to decelerate later
-			deltaX = dx;
-			deltaY = dy;
-			lastTimestamp = System.currentTimeMillis();
-			
-			e.consume();
-		}
-		
-	}
-	
-	
-	private void setViewerPopupMenu(final QuPathViewerPlus viewer) {
-		
-		final ContextMenu popup = new ContextMenu();
-		
-		MenuItem miAddRow = new MenuItem("Add row");
-		miAddRow.setOnAction(e -> viewerManager.addRow(viewer));
-		MenuItem miAddColumn = new MenuItem("Add column");
-		miAddColumn.setOnAction(e -> viewerManager.addColumn(viewer));
-		
-		MenuItem miRemoveRow = new MenuItem("Remove row");
-		miRemoveRow.setOnAction(e -> viewerManager.removeRow(viewer));
-		MenuItem miRemoveColumn = new MenuItem("Remove column");
-		miRemoveColumn.setOnAction(e -> viewerManager.removeColumn(viewer));
-
-		MenuItem miCloseViewer = new MenuItem("Close viewer");
-		miCloseViewer.setOnAction(e -> {
-			viewerManager.closeViewer(viewer);
-//				viewerManager.removeViewer(viewer);
-		});
-		MenuItem miResizeGrid = new MenuItem("Reset grid size");
-		miResizeGrid.setOnAction(e -> {
-				viewerManager.resetGridSize();
-		});
-		MenuItem miToggleSync = ActionTools.createCheckMenuItem(defaultActions.TOGGLE_SYNCHRONIZE_VIEWERS, null);
-		MenuItem miMatchResolutions = ActionTools.createMenuItem(defaultActions.MATCH_VIEWER_RESOLUTIONS);
-		Menu menuMultiview = MenuTools.createMenu(
-				"Multi-view",
-				miToggleSync,
-				miMatchResolutions,
-				miCloseViewer,
-				null,
-				miResizeGrid,
-				null,
-//				miSplitQuadrants,
-//				null,
-				miAddRow,
-				miAddColumn,
-				null,
-				miRemoveRow,
-				miRemoveColumn
-				);
-		
-		Menu menuView = MenuTools.createMenu(
-				"Display",
-				ActionTools.createCheckMenuItem(defaultActions.SHOW_ANALYSIS_PANE, null),
-				defaultActions.BRIGHTNESS_CONTRAST,
-				null,
-				ActionTools.createAction(() -> Commands.setViewerDownsample(viewer, 0.25), "400%"),
-				ActionTools.createAction(() -> Commands.setViewerDownsample(viewer, 1), "100%"),
-				ActionTools.createAction(() -> Commands.setViewerDownsample(viewer, 2), "50%"),
-				ActionTools.createAction(() -> Commands.setViewerDownsample(viewer, 10), "10%"),
-				ActionTools.createAction(() -> Commands.setViewerDownsample(viewer, 100), "1%")
-				);
-		
-		ToggleGroup groupTools = new ToggleGroup();
-		Menu menuTools = MenuTools.createMenu(
-				"Set tool",
-				ActionTools.createCheckMenuItem(defaultActions.MOVE_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.RECTANGLE_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.ELLIPSE_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.LINE_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.POLYGON_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.POLYLINE_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.BRUSH_TOOL, groupTools),
-				ActionTools.createCheckMenuItem(defaultActions.POINTS_TOOL, groupTools),
-				null,
-				ActionTools.createCheckMenuItem(defaultActions.SELECTION_MODE)
-//				ActionTools.getActionCheckBoxMenuItem(actionManager.WAND_TOOL, groupTools)
-				);
-
-		
-		// Handle awkward 'TMA core missing' option
-		CheckMenuItem miTMAValid = new CheckMenuItem("Set core valid");
-		miTMAValid.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), false));
-		CheckMenuItem miTMAMissing = new CheckMenuItem("Set core missing");
-		miTMAMissing.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), true));
-		
-		Menu menuTMA = new Menu("TMA");
-		MenuTools.addMenuItems(
-				menuTMA,
-				miTMAValid,
-				miTMAMissing,
-				null,
-				defaultActions.TMA_ADD_NOTE,
-				null,
-				MenuTools.createMenu(
-						"Add",
-					createImageDataAction(imageData -> TMACommands.promptToAddRowBeforeSelected(imageData), "Add TMA row before"),
-					createImageDataAction(imageData -> TMACommands.promptToAddRowAfterSelected(imageData), "Add TMA row after"),
-					createImageDataAction(imageData -> TMACommands.promptToAddColumnBeforeSelected(imageData), "Add TMA column before"),
-					createImageDataAction(imageData -> TMACommands.promptToAddColumnAfterSelected(imageData), "Add TMA column after")
-					),
-				MenuTools.createMenu(
-						"Remove",
-					createImageDataAction(imageData -> TMACommands.promptToDeleteTMAGridRow(imageData), "Remove TMA row"),
-					createImageDataAction(imageData -> TMACommands.promptToDeleteTMAGridColumn(imageData), "column")
-					)
-				);
-		
-		
-		// Create an empty placeholder menu
-		Menu menuSetClass = MenuTools.createMenu("Set class");
-		
-		
-//		CheckMenuItem miTMAValid = new CheckMenuItem("Set core valid");
-//		miTMAValid.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), false));
-//		CheckMenuItem miTMAMissing = new CheckMenuItem("Set core missing");
-//		miTMAMissing.setOnAction(e -> setTMACoreMissing(viewer.getHierarchy(), true));
-		
-		
-		Menu menuCells = MenuTools.createMenu(
-				"Cells",
-				ActionTools.createCheckMenuItem(defaultActions.SHOW_CELL_BOUNDARIES_AND_NUCLEI, null),
-				ActionTools.createCheckMenuItem(defaultActions.SHOW_CELL_NUCLEI, null),
-				ActionTools.createCheckMenuItem(defaultActions.SHOW_CELL_BOUNDARIES, null),
-				ActionTools.createCheckMenuItem(defaultActions.SHOW_CELL_CENTROIDS, null)
-				);
-
-		
-		
-		MenuItem miClearSelectedObjects = new MenuItem("Delete object");
-		miClearSelectedObjects.setOnAction(e -> {
-			PathObjectHierarchy hierarchy = viewer.getHierarchy();
-			if (hierarchy == null)
-				return;
-			if (hierarchy.getSelectionModel().singleSelection()) {
-				GuiTools.promptToRemoveSelectedObject(hierarchy.getSelectionModel().getSelectedObject(), hierarchy);
-			} else {
-				GuiTools.promptToClearAllSelectedObjects(viewer.getImageData());
-			}
-		});
-		
-		// Create a standard annotations menu
-		Menu menuAnnotations = GuiTools.populateAnnotationsMenu(this, new Menu("Annotations"));
-		
-		SeparatorMenuItem topSeparator = new SeparatorMenuItem();
-		popup.setOnShowing(e -> {
-			// Check if we have any cells
-			ImageData<?> imageData = viewer.getImageData();
-			if (imageData == null)
-				menuCells.setVisible(false);
-			else
-				menuCells.setVisible(!imageData.getHierarchy().getDetectionObjects().isEmpty());
-			
-			
-			// Check what to show for TMA cores or annotations
-			Collection<PathObject> selectedObjects = viewer.getAllSelectedObjects();
-			PathObject pathObject = viewer.getSelectedObject();
-			menuTMA.setVisible(false);
-			if (pathObject instanceof TMACoreObject) {
-				boolean isMissing = ((TMACoreObject)pathObject).isMissing();
-				miTMAValid.setSelected(!isMissing);
-				miTMAMissing.setSelected(isMissing);
-				menuTMA.setVisible(true);
-			}
-			
-			// Add clear objects option if we have more than one non-TMA object
-			if (imageData == null || imageData.getHierarchy().getSelectionModel().noSelection() || imageData.getHierarchy().getSelectionModel().getSelectedObject() instanceof TMACoreObject)
-				miClearSelectedObjects.setVisible(false);
-			else {
-				if (imageData.getHierarchy().getSelectionModel().singleSelection()) {
-					miClearSelectedObjects.setText("Delete object");
-					miClearSelectedObjects.setVisible(true);
-				} else {
-					miClearSelectedObjects.setText("Delete objects");
-					miClearSelectedObjects.setVisible(true);					
-				}
-			}
-			
-			boolean hasAnnotations = pathObject instanceof PathAnnotationObject || (!selectedObjects.isEmpty() && selectedObjects.stream().allMatch(p -> p.isAnnotation()));
-			
-			updateSetAnnotationPathClassMenu(menuSetClass, viewer);
-			menuAnnotations.setVisible(hasAnnotations);
-			topSeparator.setVisible(hasAnnotations || pathObject instanceof TMACoreObject);
-			// Occasionally, the newly-visible top part of a popup menu can have the wrong size?
-			popup.setWidth(popup.getPrefWidth());
-		});
-		
-		
-//		popup.add(menuClassify);
-		popup.getItems().addAll(
-				miClearSelectedObjects,
-				menuTMA,
-				menuSetClass,
-				menuAnnotations,
-				topSeparator,
-				menuMultiview,
-				menuCells,
-				menuView,
-				menuTools
-				);
-		
-		popup.setAutoHide(true);
-		
-		// Enable circle pop-up for quick classification on right-click
-		CirclePopupMenu circlePopup = new CirclePopupMenu(viewer.getView(), null);
-		viewer.getView().addEventHandler(MouseEvent.MOUSE_PRESSED, e -> {
-			if ((e.isPopupTrigger() || e.isSecondaryButtonDown()) && e.isShiftDown() && !getAvailablePathClasses().isEmpty()) {
-				circlePopup.setAnimationDuration(Duration.millis(200));
-				updateSetAnnotationPathClassMenu(circlePopup, viewer);
-				circlePopup.show(e.getScreenX(), e.getScreenY());
-				e.consume();
-				return;
-			} else if (circlePopup.isShown())
-				circlePopup.hide();
-				
-			if (e.isPopupTrigger() || e.isSecondaryButtonDown()) {
-				popup.show(viewer.getView().getScene().getWindow(), e.getScreenX(), e.getScreenY());				
-				e.consume();
-			}
-		});
-		
-//		// It's necessary to make the Window the owner, since otherwise the context menu does not disappear when clicking elsewhere on the viewer
-//		viewer.getView().setOnContextMenuRequested(e -> {
-//			popup.show(viewer.getView().getScene().getWindow(), e.getScreenX(), e.getScreenY());
-////			popup.show(viewer.getView(), e.getScreenX(), e.getScreenY());
-//		});
-			
-	}
-
-	/**
-	 * Set selected TMA cores to have the specified 'missing' status.
-	 * 
-	 * @param hierarchy
-	 * @param setToMissing
-	 */
-	private static void setTMACoreMissing(final PathObjectHierarchy hierarchy, final boolean setToMissing) {
-		if (hierarchy == null)
-			return;
-		PathObject pathObject = hierarchy.getSelectionModel().getSelectedObject();
-		List<PathObject> changed = new ArrayList<>();
-		if (pathObject instanceof TMACoreObject) {
-			TMACoreObject core = (TMACoreObject)pathObject;
-			core.setMissing(setToMissing);
-			changed.add(core);
-			// Update any other selected cores to have the same status
-			for (PathObject pathObject2 : hierarchy.getSelectionModel().getSelectedObjects()) {
-				if (pathObject2 instanceof TMACoreObject) {
-					core = (TMACoreObject)pathObject2;
-					if (core.isMissing() != setToMissing) {
-						core.setMissing(setToMissing);
-						changed.add(core);
-					}
-				}
-			}
-		}
-		if (!changed.isEmpty())
-			hierarchy.fireObjectsChangedEvent(getInstance(), changed);
-	}
-		
-	
-	/**
-	 * Update a 'set annotation class' menu for a viewer immediately prior to display
-	 * 
-	 * @param menuSetClass
-	 * @param viewer
-	 */
-	void updateSetAnnotationPathClassMenu(final Menu menuSetClass, final QuPathViewer viewer) {
-		updateSetAnnotationPathClassMenu(menuSetClass.getItems(), viewer, false);
-		menuSetClass.setVisible(!menuSetClass.getItems().isEmpty());
-	}
-	
-	
-	void updateSetAnnotationPathClassMenu(final CirclePopupMenu menuSetClass, final QuPathViewer viewer) {
-		updateSetAnnotationPathClassMenu(menuSetClass.getItems(), viewer, true);
-	}
-
-		
-	void updateSetAnnotationPathClassMenu(final ObservableList<MenuItem> menuSetClassItems, final QuPathViewer viewer, final boolean useFancyIcons) {
-		// We need a viewer and an annotation, as well as some PathClasses, otherwise we just need to ensure the menu isn't visible
-		if (viewer == null || !(viewer.getSelectedObject() instanceof PathAnnotationObject) || availablePathClasses.isEmpty()) {
-			menuSetClassItems.clear();
-			return;
-		}
-		
-		PathObject mainPathObject = viewer.getSelectedObject();
-		PathClass currentClass = mainPathObject.getPathClass();
-		
-		ToggleGroup group = new ToggleGroup();
-		List<MenuItem> itemList = new ArrayList<>();
-		RadioMenuItem selected = null;
-		for (PathClass pathClass : availablePathClasses) {
-			PathClass pathClassToSet = pathClass.getName() == null ? null : pathClass;
-			String name = pathClass.getName() == null ? "None" : pathClass.toString();
-			Action actionSetClass = new Action(name, e -> {
-				List<PathObject> changed = new ArrayList<>();
-				for (PathObject pathObject : viewer.getAllSelectedObjects()) {
-					if (!pathObject.isAnnotation() || pathObject.getPathClass() == pathClassToSet)
-						continue;
-					pathObject.setPathClass(pathClassToSet);
-					changed.add(pathObject);
-				}
-				if (!changed.isEmpty())
-					viewer.getHierarchy().fireObjectClassificationsChangedEvent(this, changed);				
-			});
-			Node shape;
-			if (useFancyIcons) {
-				Ellipse r = new Ellipse(TOOLBAR_ICON_SIZE/2.0, TOOLBAR_ICON_SIZE/2.0, TOOLBAR_ICON_SIZE, TOOLBAR_ICON_SIZE);
-				if ("None".equals(name)) {
-					r.setFill(Color.rgb(255, 255, 255, 0.75));
-					
-				}
-				else
-					r.setFill(ColorToolsFX.getCachedColor(pathClass.getColor()));
-				r.setOpacity(0.8);
-				DropShadow effect = new DropShadow(6, -3, 3, Color.GRAY);
-				r.setEffect(effect);
-				shape = r;
-			} else {
-				Rectangle r = new Rectangle(0, 0, 8, 8);
-				r.setFill("None".equals(name) ? Color.TRANSPARENT : ColorToolsFX.getCachedColor(pathClass.getColor()));
-				shape = r;
-			}
-//			actionSetClass.setGraphic(r);
-			RadioMenuItem item = ActionUtils.createRadioMenuItem(actionSetClass);
-			item.graphicProperty().unbind();
-			item.setGraphic(shape);
-			item.setToggleGroup(group);
-			itemList.add(item);
-			if (pathClassToSet == currentClass)
-				selected = item;
-		}
-		group.selectToggle(selected);
-		menuSetClassItems.setAll(itemList);
-	}
 	
 	
 	/**
@@ -3065,10 +2458,6 @@ public class QuPathGUI {
 						fileBase = path.toFile().getParentFile();
 				} catch (Exception e) {};
 			}
-//			pathOld = server.getPath();
-//			try {
-//				fileBase = new File(pathOld).getParentFile();
-//			} catch (Exception e) {};
 		}
 		// Prompt for a path, if required
 		File fileNew = null;
@@ -3121,7 +2510,6 @@ public class QuPathGUI {
 
 			UriImageSupport<BufferedImage> support = ImageServerProvider.getPreferredUriImageSupport(BufferedImage.class, pathNew);
 			List<ServerBuilder<BufferedImage>> builders = support == null ? Collections.emptyList() : support.getBuilders();
-//			List<ImageServer<BufferedImage>> serverList = ImageServerProvider.getServerList(pathNew, BufferedImage.class);
 			
 			if (builders.isEmpty()) {
 				String message = "Unable to build ImageServer for " + pathNew + ".\nSee View > Show log for more details";
@@ -3179,22 +2567,16 @@ public class QuPathGUI {
 				}
 				
 				viewer.setImageData(imageData);
-//				setInitialLocationAndMagnification(viewer);
 
 				if (imageData.getImageType() == ImageType.UNSET && PathPrefs.imageTypeSettingProperty().get() == ImageTypeSetting.PROMPT) {
 					var type = GuiTools.estimateImageType(serverNew, imageRegionStore.getThumbnail(serverNew, 0, 0, true));
 					ImageDetailsPane.promptToSetImageType(imageData, type);
 				}
 
-//				// Reset the object hierarchy to clear any ROIs etc.
-//				hierarchy.clearAll();
-//				hierarchy.getSelectionModel().resetSelection();
-				
 				return true;
 			} else {
 				// Show an error message if we can't open the file
 				Dialogs.showErrorNotification("Open image", "Sorry, I can't open " + pathNew);
-//				logger.error("Unable to build whole slide server for path '{}'", pathNew);
 			}
 		}
 		return false;
@@ -4112,45 +3494,19 @@ public class QuPathGUI {
 			previousTool = getSelectedTool();
 		this.selectedToolProperty.set(tool);
 	}
+
 	
-	
-	/**
-	 * Request that a specified Jar file be added to the extension classpath.
-	 * <p>
-	 * Note: This is really intended for dependencies that should remain where they are 
-	 * on disk (e.g. because they are included in other applications).
-	 * <p>
-	 * Jars containing QuPath extensions should be copied directly into the extensions 
-	 * directory instead.
-	 * 
-	 * @param file
-	 * @return
-	 */
-	public boolean addExtensionJar(final File file) {
-		return extensionClassLoader.addJAR(file);
-	}
-	
-	
-	
-	/**
-	 * Set the cursor for all the viewers.
-	 */
-	void updateCursor() {
+	private void updateCursorForSelectedTool() {
 		if (stage == null || stage.getScene() == null)
 			return;
 		var mode = getSelectedTool();
 		if (mode == PathTools.MOVE)
-			updateCursor(Cursor.HAND);
+			setCursorForAllViewers(Cursor.HAND);
 		else
-			updateCursor(Cursor.DEFAULT);
+			setCursorForAllViewers(Cursor.DEFAULT);
 	}
 	
-	/**
-	 * Set the cursor for all the viewers.
-	 * 
-	 * @param cursor
-	 */
-	void updateCursor(final Cursor cursor) {
+	private void setCursorForAllViewers(final Cursor cursor) {
 		for (QuPathViewer viewer : getViewers())
 			viewer.getView().setCursor(cursor);
 	}
@@ -4399,7 +3755,7 @@ public class QuPathGUI {
 		return stage;
 	}
 	
-	private void updateMagnificationString() {
+	void updateMagnificationString() {
 		if (toolbar == null)
 			return;
 		toolbar.updateMagnificationDisplay(getViewer());
@@ -4439,12 +3795,12 @@ public class QuPathGUI {
 					var versionString = getVersionString();
 					if (versionString != null)
 						name = name + " (" + versionString + ")";
-					var imageData = imageDataProperty.get();
+					var imageData = imageDataProperty().get();
 					if (imageData == null || !PathPrefs.showImageNameInTitleProperty().get())
 						return name;
 					return name + " - " + getDisplayedImageName(imageData);
 				},
-				projectProperty, imageDataProperty, PathPrefs.showImageNameInTitleProperty(), PathPrefs.maskImageNamesProperty());
+				projectProperty, imageDataProperty(), PathPrefs.showImageNameInTitleProperty(), PathPrefs.maskImageNamesProperty());
 	
 	
 	/**
@@ -4474,19 +3830,9 @@ public class QuPathGUI {
 	 * @return
 	 */
 	public ReadOnlyObjectProperty<ImageData<BufferedImage>> imageDataProperty() {
-		return imageDataProperty;
+		return viewerManager.imageDataProperty();
 	}
-	
-	
-	private void fireImageDataChangedEvent(final ImageData<BufferedImage> imageDataOld, ImageData<BufferedImage> imageDataNew) {		
 		
-		imageDataProperty.set(imageDataNew);
-		
-//		// A bit awkward, this... but make sure the extended scripting helper static class knows what's happened
-//		QPEx.setBatchImageData(imageDataNew);
-		
-	}
-	
 	
 	/**
 	 * Set the active project, triggering any necessary GUI updates.
@@ -4515,13 +3861,9 @@ public class QuPathGUI {
 				continue;
 			var imageData = viewer.getImageData();
 			if (imageData != null) {
-//				ProjectImageEntry<BufferedImage> entry = getProjectImageEntry(imageData);
-	//			if (entry != null) {
-					if (!checkSaveChanges(imageData))
-						return;
-					viewer.setImageData(null);
-	//			} else
-	//				ProjectImportImagesCommand.addSingleImageToProject(project, imageData.getServer(), null);
+				if (!checkSaveChanges(imageData))
+					return;
+				viewer.setImageData(null);
 			}
 		}
 		
@@ -4571,11 +3913,6 @@ public class QuPathGUI {
 				getAvailablePathClasses().setAll(pathClasses);
 			}
 		}
-		
-		// Ensure we have the required directories
-//		getProjectClassifierDirectory(true);
-//		getProjectScriptsDirectory(true);
-		
 		logger.info("Project set to {}", project);
 	}
 	
@@ -4659,7 +3996,7 @@ public class QuPathGUI {
 	 * @param imageData
 	 * @return true if the prompt 'succeeded' (i.e. user chose 'Yes' or 'No'), false if it was cancelled.
 	 */
-	private boolean promptToSaveChangesOrCancel(String dialogTitle, ImageData<BufferedImage> imageData) {
+	boolean promptToSaveChangesOrCancel(String dialogTitle, ImageData<BufferedImage> imageData) {
 		var project = getProject();
 		var entry = project == null ? null : project.getEntry(imageData);
 		File filePrevious = null;
@@ -4717,749 +4054,6 @@ public class QuPathGUI {
 		long tileCacheSize = Math.round(maxAvailable * (percentage / 100.0));
 		logger.info(String.format("Setting tile cache size to %.2f MB (%.1f%% max memory)", tileCacheSize/(1024.*1024.), percentage));
 		return tileCacheSize;
-	}
-
-
-
-	class MultiviewManager implements QuPathViewerListener {
-		
-		private List<QuPathViewerPlus> viewers = new ArrayList<>();
-		private SimpleObjectProperty<QuPathViewerPlus> activeViewerProperty = new SimpleObjectProperty<>();
-		
-		private SplitPaneGrid splitPaneGrid;
-		
-		private PathObject lastAnnotationObject = null;
-//		private ROI lastROI = null;
-		
-		final int borderWidth = 4;
-		final Color colorTransparent = Color.TRANSPARENT;
-		final Color colorBorder = Color.rgb(180, 0, 0, 0.5);
-		
-		final Border borderTransparent = new Border(new BorderStroke(colorTransparent, null, null, null));
-		final Border borderSelected = new Border(new BorderStroke(colorBorder, BorderStrokeStyle.SOLID, null, null));
-		
-//		private boolean aligningCores = false;
-		private BooleanProperty synchronizeViewers = new SimpleBooleanProperty(true);
-		private double lastX = Double.NaN;
-		private double lastY = Double.NaN;
-		private double lastDownsample = Double.NaN;
-		private double lastRotation = Double.NaN;
-		
-		public MultiviewManager(final QuPathViewerPlus defaultViewer) {
-			this.viewers.add(defaultViewer);
-			if (defaultViewer != null)
-				defaultViewer.addViewerListener(this);
-			setActiveViewer(defaultViewer);
-			splitPaneGrid = new SplitPaneGrid(defaultViewer.getView());
-		}
-		
-		public List<QuPathViewerPlus> getViewers() {
-			return Collections.unmodifiableList(viewers);
-		}
-		
-		/**
-		 * Match the display resolutions (downsamples) of all viewers to match the current viewer.
-		 * This uses calibrated pixel size information if available.
-		 */
-		public void matchResolutions() {
-			var viewer = getViewer();
-			var activeViewers = getViewers().stream().filter(v -> v.hasServer()).collect(Collectors.toList());
-			if (activeViewers.size() <= 1 || !viewer.hasServer())
-				return;
-			var cal = viewer.getServer().getPixelCalibration();
-			double pixelSize = cal.getAveragedPixelSize().doubleValue();
-			double downsample = viewer.getDownsampleFactor();
-			for (var temp : activeViewers) {
-				if (temp == viewer)
-					continue;
-				var cal2 = temp.getServer().getPixelCalibration();
-				double newDownsample;
-				double tempPixelSize = cal2.getAveragedPixelSize().doubleValue();
-				if (Double.isFinite(tempPixelSize) && Double.isFinite(pixelSize) && cal2.getPixelWidthUnit().equals(cal.getPixelWidthUnit()) && cal2.getPixelHeightUnit().equals(cal.getPixelHeightUnit())) {
-					newDownsample = (pixelSize / tempPixelSize) * downsample;
-				} else {
-					newDownsample = downsample;
-				}
-				temp.setDownsampleFactor(newDownsample);
-			}
-		}
-		
-		private void setActiveViewer(final QuPathViewerPlus viewer) {
-			QuPathViewerPlus previousActiveViewer = getActiveViewer();
-			if (previousActiveViewer == viewer)
-				return;
-			
-			ImageData<BufferedImage> imageDataOld = getImageData();
-			ImageData<BufferedImage> imageDataNew = viewer == null ? null : viewer.getImageData();
-			if (previousActiveViewer != null) {
-				previousActiveViewer.setBorderColor(null);
-//				activeViewer.setBorder(BorderFactory.createLineBorder(colorTransparent, borderWidth));
-//				activeViewer.setBorder(null);
-				deactivateTools(previousActiveViewer);
-				
-				// Grab reference to the current annotation, if there is one
-				PathObject pathObjectSelected = previousActiveViewer.getSelectedObject();
-				if (pathObjectSelected instanceof PathAnnotationObject) {
-					lastAnnotationObject = pathObjectSelected;					
-				}
-			}
-			this.activeViewerProperty.set(viewer);
-			lastX = Double.NaN;
-			lastY = Double.NaN;
-			lastDownsample = Double.NaN;
-			lastRotation = Double.NaN;
-			if (viewer != null) {
-//				activeViewer.getView().setBorder(null);
-				viewer.setBorderColor(colorBorder);
-//				activeViewer.setBorder(BorderFactory.createLineBorder(colorBorder, borderWidth));
-				activateTools(viewer);
-//				QuPathGUI qupath = QuPathGUI.this; // New to me... http://stackoverflow.com/questions/1816458/getting-hold-of-the-outer-class-object-from-the-inner-class-object
-//				if (qupath != null)
-//					qupath.imageDataChanged(null, imageDataOld, imageDataNew);
-				
-				if (viewer.getServer() != null) {
-					lastX = viewer.getCenterPixelX();
-					lastY = viewer.getCenterPixelY();
-					lastDownsample = viewer.getDownsampleFactor();
-					lastRotation = viewer.getRotation();
-				}
-				
-				updateMagnificationString();
-
-			}
-			logger.debug("Active viewer set to {}", viewer);
-			fireImageDataChangedEvent(imageDataOld, imageDataNew);
-		}
-		
-		public QuPathViewerPlus getActiveViewer() {
-			return activeViewerProperty.get();
-		}
-		
-		public ReadOnlyObjectProperty<QuPathViewerPlus> activeViewerProperty() {
-			return activeViewerProperty;
-		}
-		
-		public Node getNode() {
-			return splitPaneGrid.getMainSplitPane();
-		}
-		
-		
-		
-		/**
-		 * Create a viewer, adding it to the stored array but not adding it to any component (which is left up to the calling code to handle)
-		 * @return
-		 */
-		protected QuPathViewerPlus createViewer() {
-			QuPathViewerPlus viewerNew = new QuPathViewerPlus(null, imageRegionStore, overlayOptions, viewerDisplayOptions);
-			setupViewer(viewerNew);
-			viewerNew.addViewerListener(this);
-			viewers.add(viewerNew);
-			return viewerNew;
-		}
-		
-		
-		SplitPane getAncestorSplitPane(Node node) {
-			while (node != null && !(node instanceof SplitPane))
-				node = node.getParent();
-			return (SplitPane)node;
-		}
-		
-		/**
-		 * Try to remove the row containing the specified viewer, notifying the user if this isn't possible.
-		 * @param viewer
-		 * @return true if the row was removed, false otherwise
-		 */
-		public boolean removeRow(final QuPathViewer viewer) {
-//			if (viewer.getServer() != null)
-//				System.err.println(viewer.getServer().getShortServerName());
-			// Note: These are the internal row numbers... these don't necessarily match with the displayed row (?)
-			int row = splitPaneGrid.getRow(viewer.getView());
- 			if (row < 0) {
-				// Shouldn't occur...
-				Dialogs.showErrorMessage("Multiview error", "Cannot find " + viewer + " in the grid!");
-				return false;
-			}
- 			if (splitPaneGrid.nRows() == 1) {
-				Dialogs.showErrorMessage("Close row error", "The last row can't be removed!");
-				return false;
- 			}
- 			
-			int nOpen = splitPaneGrid.countOpenViewersForRow(row);
-			if (nOpen > 0) {
-				Dialogs.showErrorMessage("Close row error", "Please close all open viewers in selected row, then try again");
-//				DisplayHelpers.showErrorMessage("Close row error", "Please close all open viewers in row " + row + ", then try again");
-				return false;
-			}
-			splitPaneGrid.removeRow(row);
-			splitPaneGrid.resetGridSize();
-			// Make sure the viewer list is up-to-date
-			refreshViewerList();
-			return true;
-		}
-		
-		
-		/**
-		 * Check all viewers to see if they are associated with a scene, and remove them from the list if not.
-		 */
-		private void refreshViewerList() {
-			// Remove viewers from the list if they aren't associated with anything
-			// Easiest way is to check for a scene
-			Iterator<? extends QuPathViewer> iter = viewers.iterator();
-			while (iter.hasNext()) {
-				if (iter.next().getView().getScene() == null)
-					iter.remove();
-			}
-		}
-		
-		
-		/**
-		 * Close the image within a viewer, prompting to save changes if necessary.
-		 * 
-		 * @param viewer
-		 * @return True if the viewer no longer contains an open image (either because it never did contain one, or 
-		 * the image was successfully closed), false otherwise (e.g. if the user thwarted the close request)
-		 */
-		public boolean closeViewer(final QuPathViewer viewer) {
-			return closeViewer("Save changes", viewer);
-		}
-		
-		/**
-		 * Close the image within a viewer, prompting to save changes if necessary.
-		 * 
-		 * @param dialogTitle Name to use within any displayed dialog box.
-		 * @param viewer
-		 * @return True if the viewer no longer contains an open image (either because it never did contain one, or 
-		 * the image was successfully closed), false otherwise (e.g. if the user thwarted the close request)
-		 */
-		public boolean closeViewer(final String dialogTitle, final QuPathViewer viewer) {
-			ImageData<BufferedImage> imageData = viewer.getImageData();
-			if (imageData == null)
-				return true;
-			// Deal with saving, if necessary
-			if (imageData.isChanged()) {
-				if (!isReadOnly() && !promptToSaveChangesOrCancel(dialogTitle, imageData))
-					return false;
-			}
-			viewer.setImageData(null);
-			return true;
-		}
-		
-		
-		/**
-		 * Try to remove the column containing the specified viewer, notifying the user if this isn't possible.
-		 * @param viewer
-		 * @return true if the column was removed, false otherwise
-		 */
-		public boolean removeColumn(final QuPathViewer viewer) {
-			int col = splitPaneGrid.getColumn(viewer.getView());
-			if (col < 0) {
-				// Shouldn't occur...
-				Dialogs.showErrorMessage("Multiview error", "Cannot find " + viewer + " in the grid!");
-				return false;
-			}
-			
- 			if (splitPaneGrid.nCols() == 1) {
-				Dialogs.showErrorMessage("Close row error", "The last row can't be removed!");
-				return false;
- 			}
-			
-			int nOpen = splitPaneGrid.countOpenViewersForColumn(col);
-			if (nOpen > 0) {
-				Dialogs.showErrorMessage("Close column error", "Please close all open viewers in selected column, then try again");
-//				DisplayHelpers.showErrorMessage("Close column error", "Please close all open viewers in column " + col + ", then try again");
-				return false;
-			}
-			splitPaneGrid.removeColumn(col);
-			splitPaneGrid.resetGridSize();
-			// Make sure the viewer list is up-to-date
-			refreshViewerList();
-			return true;
-		}
-		
-		
-		public void addRow(final QuPathViewer viewer) {
-			splitViewer(viewer, false);
-			splitPaneGrid.resetGridSize();
-		}
-
-		public void addColumn(final QuPathViewer viewer) {
-			splitViewer(viewer, true);
-			splitPaneGrid.resetGridSize();
-		}
-
-		
-		public void splitViewer(final QuPathViewer viewer, final boolean splitVertical) {
-			if (!viewers.contains(viewer))
-				return;
-			
-			if (splitVertical) {
-				splitPaneGrid.addColumn(splitPaneGrid.getColumn(viewer.getView()));
-			} else {
-				splitPaneGrid.addRow(splitPaneGrid.getRow(viewer.getView()));
-			}
-		}
-		
-		public void resetGridSize() {
-			splitPaneGrid.resetGridSize();
-		}
-		
-		
-		public void repaintViewers() {
-			for (QuPathViewer v : viewers)
-				v.repaint();
-		}
-
-		@Override
-		public void imageDataChanged(QuPathViewer viewer, ImageData<BufferedImage> imageDataOld, ImageData<BufferedImage> imageDataNew) {
-			if (viewer != null && viewer == getActiveViewer()) {
-				if (viewer.getServer() != null) {
-					// Setting these to NaN prevents unexpected jumping when a new image is opened
-					lastX = Double.NaN;
-					lastY = Double.NaN;
-					lastDownsample = Double.NaN;
-					lastRotation = Double.NaN;
-				}
-				fireImageDataChangedEvent(imageDataOld, viewer.getImageData());
-			}
-		}
-
-		@Override
-		public void visibleRegionChanged(QuPathViewer viewer, Shape shape) {
-			if (viewer == null)
-				return;
-			if (viewer != getActiveViewer() || viewer.isImageDataChanging() || zoomToFit.get()) {
-//				// Only change downsamples for non-active viewer
-//				double downsample = viewer.getDownsampleFactor();
-//				if (synchronizeViewers) {
-//					for (QuPathViewer v : viewers) {
-//						double oldDownsample = v.getDownsampleFactor();
-//						if (!GeneralTools.almostTheSame(downsample, oldDownsample, 0.0001)) {
-//							v.setDownsampleFactor(downsample);
-//						}
-//					}
-//				}
-				return;
-			} else {
-				// Update magnification info
-				updateMagnificationString();
-			}
-			
-			QuPathViewerPlus activeViewer = getActiveViewer();
-			double x = activeViewer.getCenterPixelX();
-			double y = activeViewer.getCenterPixelY();
-			double rotation = activeViewer.getRotation();
-			double dx = Double.NaN, dy = Double.NaN, dr = Double.NaN;
-			
-			double downsample = viewer.getDownsampleFactor();
-			double relativeDownsample = viewer.getDownsampleFactor() / lastDownsample;
-			
-			// Shift as required, assuming we aren't aligning cores
-//			if (!aligningCores) {
-//			synchronizeViewers = true;
-			if (synchronizeViewers.get()) {
-				if (!Double.isNaN(lastX + lastY)) {
-					dx = x - lastX;
-					dy = y - lastY;
-					dr = rotation - lastRotation;
-				}
-				
-				for (QuPathViewer v : viewers) {
-					if (v == viewer)
-						continue;
-					if (!Double.isNaN(relativeDownsample))
-						v.setDownsampleFactor(v.getDownsampleFactor() * relativeDownsample, -1, -1, false);
-					if (!Double.isNaN(dr) && dr != 0)
-						v.setRotation(v.getRotation() + dr);
-					
-					// Shift as required - correcting for rotation
-					double downsampleRatio = v.getDownsampleFactor() / downsample;
-					if (!Double.isNaN(dx) && !Double.isNaN(downsampleRatio)) {
-						
-						double rot = rotation - v.getRotation();
-						double sin = Math.sin(rot);
-						double cos = Math.cos(rot);
-						
-						double dx2 = dx * downsampleRatio;
-						double dy2 = dy * downsampleRatio;
-
-						double dx3 = cos * dx2 - sin * dy2;
-						double dy3 = sin * dx2 + cos * dy2;
-						
-						v.setCenterPixelLocation(v.getCenterPixelX() + dx3, v.getCenterPixelY() + dy3);
-					}
-				}
-			}
-			
-			lastX = x;
-			lastY = y;
-			lastDownsample = downsample;
-			lastRotation = rotation;
-		}
-		
-		
-		public boolean getSynchronizeViewers() {
-			return synchronizeViewers.get();
-		}
-		
-		public void setSynchronizeViewers(final boolean synchronizeViewers) {
-			this.synchronizeViewers.set(synchronizeViewers);
-		}
-		
-		public ReadOnlyBooleanProperty synchronizeViewersProperty() {
-			return synchronizeViewers;
-		}
-		
-
-		@Override
-		public void selectedObjectChanged(QuPathViewer viewer, PathObject pathObjectSelected) {
-			// Store any annotation ROIs, which might need to be transferred
-			if (pathObjectSelected instanceof PathAnnotationObject) {
-				lastAnnotationObject = pathObjectSelected;
-				return;
-			}
-			
-			// Don't handle unselected viewers
-			if (viewer != getActiveViewer()) {
-				return;
-			}
-			
-			// Synchronize TMA cores
-			if (!(pathObjectSelected instanceof TMACoreObject))
-				return;
-			
-			// Thwart the upcoming region shift
-			lastX = Double.NaN;
-			lastY = Double.NaN;
-			lastDownsample = Double.NaN;
-			lastRotation = Double.NaN;
-			
-//			aligningCores = true;
-			String coreName = ((TMACoreObject)pathObjectSelected).getName();
-			for (QuPathViewer v : viewers) {
-				if (v == viewer)
-					continue;
-				PathObjectHierarchy hierarchy = v.getHierarchy();
-				if (hierarchy == null || hierarchy.getTMAGrid() == null)
-					continue;
-				
-				TMAGrid tmaGrid = hierarchy.getTMAGrid();
-				TMACoreObject core = tmaGrid.getTMACore(coreName);
-				if (core != null) {
-					v.setSelectedObject(core);
-					double cx = core.getROI().getCentroidX();
-					double cy = core.getROI().getCentroidY();
-					v.setCenterPixelLocation(cx, cy);
-				}
-			}
-		}
-		
-		
-		
-		public boolean applyLastAnnotationToActiveViewer() {
-			if (lastAnnotationObject == null) {
-				logger.info("No annotation object to copy");
-				return false;
-			}
-			
-			QuPathViewerPlus activeViewer = getActiveViewer();
-			if (activeViewer == null || activeViewer.getHierarchy() == null) {
-				logger.info("No active viewer available");
-				return false;
-			}
-			
-			PathObjectHierarchy hierarchy = activeViewer.getHierarchy();
-			if (PathObjectTools.hierarchyContainsObject(hierarchy, lastAnnotationObject)) {
-				logger.info("Hierarchy already contains annotation object!");
-				return false;
-			}
-			
-			ROI roi = lastAnnotationObject.getROI().duplicate();
-			
-			// If we are within a TMA core, try to apply any required translations
-			TMACoreObject coreNewParent = null;
-			if (hierarchy.getTMAGrid() != null) {
-				TMACoreObject coreParent = null;
-				PathObject parent = lastAnnotationObject.getParent();
-				while (parent != null) {
-					if (parent instanceof TMACoreObject) {
-						coreParent = (TMACoreObject)parent;
-						break;
-					} else
-						parent = parent.getParent();
-				}
-				if (coreParent != null) {
-					coreNewParent = hierarchy.getTMAGrid().getTMACore(coreParent.getName());
-					if (coreNewParent != null) {
-						double rotation = activeViewer.getRotation();
-//						if (rotation == 0) {
-							double dx = coreNewParent.getROI().getCentroidX() - coreParent.getROI().getCentroidX();
-							double dy = coreNewParent.getROI().getCentroidY() - coreParent.getROI().getCentroidY();
-							roi = roi.translate(dx, dy);
-							// TODO: Deal with rotations... it's a bit tricky...
-//						} else {
-						// TODO: Check how best to handle transferring ROIs with rotation involved
-						if (rotation != 0) {
-							AffineTransform transform = new AffineTransform();
-							transform.rotate(-rotation, coreNewParent.getROI().getCentroidX(), coreNewParent.getROI().getCentroidY());
-							logger.info("ROTATING: " + transform);
-							Area area = RoiTools.getArea(roi);
-							area.transform(transform);
-							roi = RoiTools.getShapeROI(area, roi.getImagePlane());
-						}
-					}
-				}
-			}
-			
-			
-			PathObject annotation = PathObjects.createAnnotationObject(roi, lastAnnotationObject.getPathClass());
-//			hierarchy.addPathObject(annotation, false);
-			
-//			// Make sure any core parent is set
-			hierarchy.addObjectBelowParent(coreNewParent, annotation, true);
-			
-			activeViewer.setSelectedObject(annotation);
-			return true;
-		}
-
-		@Override
-		public void viewerClosed(QuPathViewer viewer) {}
-
-		public QuPathViewerPlus getViewer() {
-			return getActiveViewer();
-		}
-		
-		
-		
-		
-		
-		
-		
-		class SplitPaneGrid {
-			
-			private SplitPane splitPaneMain = new SplitPane();
-			private List<SplitPane> splitPaneRows = new ArrayList<>();
-			
-			SplitPaneGrid(final Node node) {
-				splitPaneMain.setOrientation(Orientation.VERTICAL);
-				SplitPane splitRow = new SplitPane();
-				splitRow.setOrientation(Orientation.HORIZONTAL);
-				splitRow.getItems().add(node);
-				splitPaneRows.add(splitRow);
-				splitPaneMain.getItems().add(splitRow);
-			}
-			
-			SplitPane getMainSplitPane() {
-				return splitPaneMain;
-			}
-			
-			
-			void addRow(final int position) {
-				SplitPane splitRow = new SplitPane();
-				splitRow.setOrientation(Orientation.HORIZONTAL);
-				
-				// For now, we create a row with the same number of columns in every row
-				// Create viewers & bind dividers
-				splitRow.getItems().clear();
-				SplitPane firstRow = splitPaneRows.get(0);
-				splitRow.getItems().add(createViewer().getView());
-				for (int i = 0; i < firstRow.getDividers().size(); i++) {
-					splitRow.getItems().add(createViewer().getView());
-//					splitRow.getDividers().get(i).positionProperty().bindBidirectional(firstRow.getDividers().get(i).positionProperty());
-				}
-				
-				// Ensure the new divider takes up half the space
-				double lastDividerPosition = position == 0 ? 0 : splitPaneMain.getDividers().get(position-1).getPosition();
-				double nextDividerPosition = position >= splitPaneRows.size()-1 ? 1 : splitPaneMain.getDividers().get(position).getPosition();
-				splitPaneRows.add(position, splitRow);
-				splitPaneMain.getItems().add(position+1, splitRow);
-				splitPaneMain.setDividerPosition(position, (lastDividerPosition + nextDividerPosition)/2);
-				
-				refreshDividerBindings();
-			}
-			
-			
-			
-			boolean removeRow(final int row) {
-				if (row < 0 || row >= splitPaneRows.size() || splitPaneRows.size() == 1) {
-					logger.error("Cannot remove row {} from grid with {} rows", row, splitPaneRows.size());
-					return false;
-				}
-				SplitPane splitPane = splitPaneRows.remove(row);
-//				// WeakHashMap should take care of this... but check anyway
-//				for (Node node : splitPane.getItems())
-//					viewerMap.remove(node);
-				splitPaneMain.getItems().remove(splitPane);
-				refreshDividerBindings();
-				return true;
-			}
-			
-			
-			/**
-			 * Restore all grid panels to be the same size
-			 */
-			public void resetGridSize() {
-				resetDividers(splitPaneRows.get(0)); // Because of property binding, this should be enough
-				resetDividers(splitPaneMain);
-			}
-			
-			
-			void resetDividers(final SplitPane splitPane) {
-				int n = splitPane.getItems().size();
-				if (n <= 1)
-					return;
-				if (n == 2) {
-					splitPane.setDividerPosition(0, 0.5);
-					return;
-				}
-				double[] positions = new double[n-1];
-				for (int i = 0; i < positions.length; i++)
-					positions[i] = (i + 1.0) / (double)n;
-				splitPane.setDividerPositions(positions);
-			}
-
-			
-			boolean removeColumn(final int col) {
-				if (col < 0 || col >= nCols() || nCols() == 1) {
-					logger.error("Cannot remove column {} from grid with {} columns", col, nCols());
-					return false;
-				}
-				for (SplitPane splitRow : splitPaneRows) {
-					splitRow.getItems().remove(col);
-				}
-				refreshDividerBindings();
-				return true;
-			}
-			
-			
-			int countOpenViewersForRow(final int row) {
-				int count = 0;
-				for (QuPathViewer viewer : getViewers()) {
-					if (row == getRow(viewer.getView()) && viewer.hasServer())
-						count++;
-				}
-				return count;
-			}
-			
-			
-			int countOpenViewersForColumn(final int col) {
-				int count = 0;
-				for (QuPathViewer viewer : getViewers()) {
-					if (col == getColumn(viewer.getView()) && viewer.hasServer())
-						count++;
-				}
-				return count;				
-			}
-
-			
-			/**
-			 * Update all divider bindings so they match the first row
-			 */
-			void refreshDividerBindings() {
-				SplitPane firstRow = splitPaneRows.get(0);
-				for (int r = 1; r < splitPaneRows.size(); r++) {
-					SplitPane splitRow = splitPaneRows.get(r);
-					for (int c = 0; c < splitRow.getDividers().size(); c++) {
-						splitRow.getDividers().get(c).positionProperty().bindBidirectional(firstRow.getDividers().get(c).positionProperty());
-					}
-				}
-			}
-			
-			
-			void addColumn(final int position) {
-				SplitPane firstRow = splitPaneRows.get(0);
-				double lastDividerPosition = position == 0 ? 0 : firstRow.getDividers().get(position-1).getPosition();
-				double nextDividerPosition = position >= firstRow.getItems().size()-1 ? 1 : firstRow.getDividers().get(position).getPosition();
-				
-				firstRow.getItems().add(position+1, createViewer().getView());
-				Divider firstDivider = firstRow.getDividers().get(position);
-				firstDivider.setPosition((lastDividerPosition + nextDividerPosition)/2);
-				for (int i = 1; i < splitPaneRows.size(); i++) {
-					SplitPane splitRow = splitPaneRows.get(i);
-					splitRow.getItems().add(position+1, createViewer().getView());
-				}
-				
-				refreshDividerBindings();
-			}
-			
-			
-			public int getRow(final Node node) {
-				int count = 0;
-				for (SplitPane row : splitPaneRows) {
-					int ind = row.getItems().indexOf(node);
-					if (ind >= 0)
-						return count;
-					count++;
-				}
-				return -1;
-			}
-
-			public int getColumn(final Node node) {
-				for (SplitPane row : splitPaneRows) {
-					int ind = row.getItems().indexOf(node);
-					if (ind >= 0)
-						return ind;
-				}
-				return -1;
-			}
-
-			public int nRows() {
-				return splitPaneRows.size();
-			}
-
-			public int nCols() {
-				return splitPaneRows.get(0).getDividers().size() + 1;
-			}
-			
-		}
-		
-		
-	}
-	
-	
-	class QuPathUncaughtExceptionHandler implements UncaughtExceptionHandler {
-		
-		private long lastExceptionTimestamp = 0L;
-		private String lastExceptionMessage = null;
-		
-		private long sameExceptionCount = 0;
-		private long minDelay = 1000;
-		
-		@Override
-		public void uncaughtException(Thread t, Throwable e) {
-			// Avoid showing the same message repeatedly
-			String msg = e.getLocalizedMessage();
-			long timestamp = System.currentTimeMillis();
-			try {
-				if (timestamp - lastExceptionTimestamp < minDelay && 
-						Objects.equals(msg, lastExceptionMessage)) {
-					sameExceptionCount++;
-					// Don't continually log the full stack trace
-					if (sameExceptionCount > 3)
-						logger.error("{} (see full stack trace above, or use 'debug' log level)", e.getLocalizedMessage());
-					else
-						logger.debug(e.getLocalizedMessage(), e);
-					return;
-				} else
-					sameExceptionCount = 0;
-	
-				if (e instanceof OutOfMemoryError) {
-					// Try to reclaim any memory we can
-					getViewer().getImageRegionStore().clearCache(true, true);
-					Dialogs.showErrorNotification("Out of memory error",
-							"Out of memory! You may need to decrease the 'Number of parallel threads' in the preferences, "
-							+ "then restart QuPath.");
-					logger.error(e.getLocalizedMessage(), e);
-				} else {
-					Dialogs.showErrorNotification("QuPath exception", e);
-					if (defaultActions.SHOW_LOG != null)
-						defaultActions.SHOW_LOG.handle(null);
-				}
-			} finally {
-				lastExceptionMessage = msg;
-				lastExceptionTimestamp = timestamp;				
-			}
-		}
-		
 	}
